Address skill install review findings
Document and enforce the exact-one install source contract across the
web API, browser client, and built-in skill install tool.

Require uploaded skill bundles to include a `.skill` filename, keep the
web test fixture directory alive for the fixture lifetime, and move
install parsing helpers out of the web skills handler so the handler
stays within the file-size limit.

Allow top-level `oneOf` in the strict schema validator so tool schemas
can express mutually exclusive source fields.
